move typing file to be in codes instead of codes/plasmod
